rag_service/requirements.txt
pymssql==2.2.7
flask==2.0.1
flask-cors==3.0.10
pyjwt==2.3.0
langchain==0.0.1
chromadb==0.1.0